+++ activerecord/lib/active_record/relation/query_methods.rb
@@ -679,11 +679,21 @@ module ActiveRecord
     #   #     WHEN "users"."id" = 3 THEN 3
     #   #   END ASC
     #
+    # To group values together, an array can be passed as a value.
+    #
+    #   User.in_order_of(:id, [[1, 5], 3])
+    #   # SELECT "users".* FROM "users"
+    #   #   WHERE "users"."id" IN (1, 5, 3)
+    #   #   ORDER BY CASE
+    #   #     WHEN "users"."id" IN (1, 5) THEN 1
+    #   #     WHEN "users"."id" = 3 THEN 2
+    #   #   END ASC
+    #
     # +column+ can point to an enum column; the actual query generated may be different depending
     # on the database adapter and the column definition.
     #
     #   class Conversation < ActiveRecord::Base
-    #     enum :status, [ :active, :archived ]
+    #     enum :status, [ :active, :archived, :draft ]
     #   end
     #
     #   Conversation.in_order_of(:status, [:archived, :active])
@@ -692,6 +702,16 @@ module ActiveRecord
     #   #   ORDER BY CASE
     #   #     WHEN "conversations"."status" = 1 THEN 1
     #   #     WHEN "conversations"."status" = 0 THEN 2
+    #   #   END ASC
+    #
+    # For an array value each individual value of the array is typecasted.
+    #
+    #   Conversation.in_order_of(:status, [[:archived, :active], :draft])
+    #   # SELECT "conversations".* FROM "conversations"
+    #   #   WHERE "conversations"."status" IN (2, 1, 0)
+    #   #   ORDER BY CASE
+    #   #     WHEN "conversations"."status" IN (1, 0) THEN 1
+    #   #     WHEN "conversations"."status" = 2 THEN 2
     #   #   END ASC
     #
     # +values+ can also include +nil+.
@@ -721,12 +741,22 @@ module ActiveRecord
       references = column_references([column])
       self.references_values |= references unless references.empty?
 
-      values = values.map { |value| model.type_caster.type_cast_for_database(column, value) }
+      values = values.map do |value|
+        if value.is_a?(Array)
+          value.map do |current_value|
+            model.type_caster.type_cast_for_database(column, current_value)
+          end
+        else
+          model.type_caster.type_cast_for_database(column, value)
+        end
+      end
       arel_column = column.is_a?(Arel::Nodes::SqlLiteral) ? column : order_column(column.to_s)
 
       scope = spawn.order!(build_case_for_value_position(arel_column, values, filter: filter))
 
       if filter
+        values = values.flatten(1)
+
         where_clause =
           if values.include?(nil)
             arel_column.in(values.compact).or(arel_column.eq(nil))
@@ -2174,7 +2204,15 @@ module ActiveRecord
       def build_case_for_value_position(column, values, filter: true)
         node = Arel::Nodes::Case.new
         values.each.with_index(1) do |value, order|
-          node.when(column.eq(value)).then(order)
+          if value.is_a?(Array)
+            if value.include?(nil)
+              node.when(column.in(value.compact).or(column.eq(nil))).then(order)
+            else
+              node.when(column.in(value)).then(order)
+            end
+          else
+            node.when(column.eq(value)).then(order)
+          end
         end
 
         node = node.else(values.length + 1) unless filter